A first very simple and stupid implementation of the reachability graph, calculated by the GPU. Thus the naming is confusing, but this will be changed later. First the algorithm checks, whether the net is covered by a place-invariant, to guarantee it is structurally bounded. There are two versions of the kernel: CoverabilityGraph_2D.cl and -3D.cl. This is for a test, whether it's better to loop over the places in the marking (2D) or to expand this loop to another dimension (3D). So far, the 2D approach seems faster.

+++ src/uniol/apt/gpu/plain/IsPlain.cl
@@ -1,6 +1,6 @@
 // checks whether a given petri net is plain.
 // result is initialized true
-kernel void IsPlain(global const int* weights, global int8* result, int count) {
+kernel void IsPlain(global const int* weights, global int* result, int count) {
 	// get index into global data array
 	int gid = get_global_id(0);
